backend/requirements.txt
flask
flask-cors
pyjwt
google-cloud-storage
firebase-admin
selenium
webdriver-manager
pytest